fixed withdraw not allowing floating point properly

+++ src/main/java/rocks/zipcode/atm/CashMachine.java
@@ -38,7 +38,7 @@ public class CashMachine {
         }
     }
 
-    public void withdraw(int amount) {
+    public void withdraw(float amount) {
         if (accountData != null) {
             tryCall(
                     () -> bank.withdraw(accountData, amount),
